% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CASboxplots}
\alias{CASboxplots}
\title{Visualize ADAR CAS with boxplots}
\usage{
CASboxplots(data, by = "", color = "", adar =  c("ADAR1", "ADAR2"))
}
\arguments{
\item{data}{Either a \code{Seurat} object,
\code{\link[SingleCellExperiment]{SingleCellExperiment}} or
\code{\link[SpatialExperiment]{SpatialExperiment}}}

\item{by}{Character string indicating the name of the categorical variable
present in the metadata to be used to divide the samples into the boxplots
(e.g. cluster, condition, group)}

\item{color}{Character string indicating the name of the variable in the
metadata to use for coloring boxplots}

\item{adar}{Character string specifying the CAS to visualize, between ADAR1
and ADAR2}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
The function generates a boxplot showing the distribution of ADAR CAS in the
cells or spots of a \code{Seurat} or \linkS4class{SummarizedExperiment}
object across user-specified categories.
}
\examples{
data(example_sce)
CASboxplots(example_sce, by = "Group", color = "Group", adar = "ADAR1")

}
