% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CASdittoplot}
\alias{CASdittoplot}
\title{Dimensional visualization of ADAR CAS}
\usage{
CASdittoplot(data, reduction = "", adar =  c("ADAR1", "ADAR2"))
}
\arguments{
\item{data}{Either a \code{Seurat} object,
\code{\link[SingleCellExperiment]{SingleCellExperiment}} or
\code{\link[SpatialExperiment]{SpatialExperiment}}}

\item{reduction}{Character string indicating the name of a dimensionality
reduction (e.g. umap, pca) slot within the object}

\item{adar}{Character string specifying the CAS to visualize, between ADAR1
and ADAR2}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
The function shows the distribution of ADAR CAS on a dimensional reduction
plot using \code{dittoDimPlot}.
}
\examples{
data(example_sce)
CASdittoplot(example_sce, reduction = "PCA", adar = "ADAR1")

}
