% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_genes}
\alias{example_genes}
\title{Example gene list for H.sapiens}
\format{
An object of class \code{character} of length 18.
}
\usage{
data(example_genes)
}
\description{
List of genes generated with example purposes to show the functionalities of
the function \code{\link[ADARcas]{yourSignature}}.
}
\keyword{datasets}
