% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_matrix}
\alias{example_matrix}
\title{Example expression data for H. sapiens}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 53 rows and 12 columns.
}
\usage{
data(example_matrix)
}
\description{
This is an example of bulk RNA-Seq dataset with normalized gene counts of 12
samples of human embryonic stem cells (hESC h9) differentiated to forebrain
and motor neurons. RNA was collected at different time points (day 0, 10, 17
and 35) and half of the samples underwent the knockdown of ADAR1.
The dataset is publicly available with the accession number GSE56152 on GEO
(Chen et al, 2015, Cell Res).
After the normalization, the matrix was filtered to keep only 100 genes,
included those of interest (i.e. ADAR and ADARB1).
This dataset should be used only with example purpose.
}
\keyword{datasets}
