% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_matrix_Mm}
\alias{example_matrix_Mm}
\title{Example expression data for M. musculus}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 32 rows and 15 columns.
}
\usage{
data(example_matrix_Mm)
}
\description{
This is an example of bulk RNA-Seq dataset generated by the union of two
datasets of mouse neurons and brains, publicly available on GEO with the
accession numbers GSE126629 (Sapiro, 2020, Cell Rep) and GSE132214 (Chalk,
2019, Genome Biol). The first dataset consists of WT samples and samples with
Adar1, Adar2 and Zfr KD of mouse primary neurons; the second is composed of
three samples with Adar2 KO of mouse brains. After the normalization, the
matrix was filtered to keep only 100 genes, included those of interest (i.e.
Adar and Adarb1).
This dataset should be used only with example purpose.
}
\keyword{datasets}
