% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_network}
\alias{example_network}
\title{Example network for H. sapiens}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 72 rows and 72 columns.
}
\usage{
data(example_network)
}
\description{
This is an example of gene regulatory network reconstructed with the R
package \pkg{parmigene} to estimate the mutual information, and then by using
the additive model of ARACNe (Margolin et al, 2006, BMC Bioinformatics) to
obtain a square-weighted adjacency matrix. This network was obtained from the
gene expression matrix of the dataset with GEO accession number GSE56152.
Then, with example purpose, it was filtered to keep only 50 genes, included
those of interest (i.e. ADAR and ADARB1).
}
\keyword{datasets}
