% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_sce}
\alias{example_sce}
\title{Example SingleCellExperiment}
\format{
An object of class \code{SingleCellExperiment} with 100 rows and 100 columns.
}
\usage{
data(example_sce)
}
\description{
This is an example of a SingleCellExperiment created to test package
functions with the package \pkg{splatter}.
}
\keyword{datasets}
