% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{matrix}
\alias{matrix}
\title{Expression matrix for H. sapiens}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 13822 rows and 12 columns.
}
\usage{
data(matrix)
}
\description{
This is an example of bulk RNA-Seq dataset with normalized gene counts of 12
samples of human embryonic stem cells (hESC h9) differentiated to forebrain
and motor neurons. RNA was collected at different time points (day 0, 10, 17
and 35) and half of the samples underwent the knockdown of ADAR1.
The dataset is publicly available with the accession number GSE56152 on GEO
(Chen et al, 2015, Cell Res).
}
\keyword{datasets}
