% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{neuronalMmSignature}
\alias{neuronalMmSignature}
\title{Compute ADAR activity with the mouse-specific neuronal signature}
\usage{
neuronalMmSignature(data, object)
}
\arguments{
\item{data}{Either a numeric expression matrix with genes as rows and samples
as columns or an ExpressionSet from a \code{Seurat} object,
\code{\link[SingleCellExperiment]{SingleCellExperiment}} or
\code{\link[SpatialExperiment]{SpatialExperiment}}}

\item{object}{Character string indicating the type of object, either bulk or
eset}
}
\value{
Matrix with ADAR activity scores or eset object updated
}
\description{
The function returns ADAR activity scores for each sample/spot/cell depending
on the starting expression matrix given as input. It employs a signature
derived from a dataset of neurons and whole brain of mice, which is
ideal to measure ADAR activity in a mouse neuronal context.
}
\examples{
data(mouse_neuronal_regulon)
data(example_matrix_Mm)
res <- neuronalMmSignature(example_matrix_Mm, object = "bulk")

}
