% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tumoralSignature}
\alias{tumoralSignature}
\title{Compute ADAR activity with the cancer-related signature}
\usage{
tumoralSignature(data, object)
}
\arguments{
\item{data}{Either a numeric expression matrix with genes as rows and samples
as columns or an ExpressionSet from a \code{Seurat} object,
\code{\link[SingleCellExperiment]{SingleCellExperiment}} or
\code{\link[SpatialExperiment]{SpatialExperiment}}}

\item{object}{Character string indicating the type of object, either bulk or
eset}
}
\value{
Matrix with ADAR activity scores or eset object updated
}
\description{
The function returns ADAR activity scores for each sample/spot/cell depending
on the starting expression matrix given as input. It employs a signature
derived from the union of two datasets of A549 cell line samples, which is
ideal to measure ADAR activity in a cancer-related context.
}
\examples{
data(tumoral_regulon)
data(example_matrix)
res <- tumoralSignature(example_matrix, object = "bulk")

}
